requirements.txt
face_recognition==1.3.0
Flask==3.1.1
imutils==0.5.4
numpy==2.3.0
Pillow==11.2.1
Requests==2.32.4
tqdm==4.67.1
